Fixes bugs of text size

+++ PowerPointLabs/PowerPointLabs/CaptionsFormat.cs
@@ -11,7 +11,7 @@ namespace PowerPointLabs
     {
         public static MsoTextEffectAlignment defaultAlignment = MsoTextEffectAlignment.msoTextEffectAlignmentCentered;
         public static Color defaultColor = Color.White;
-        public static int defaultSize = 12;
+        public static float defaultSize = 12;
         public static bool defaultBold = false;
         public static bool defaultItalic = false;
         public static Color defaultFillColor = Color.Black;
